Moved privileges to separeted file
