+++ api/model/database/privileges.py
@@ -1,0 +1,9 @@
+from api import db
+
+class Privilegio(db.Model):
+    __tablename__ = 'privilegios'
+    id = db.Column(db.Integer, primary_key=True)
+    user_type = db.Column(db.String(80), unique=True, nullable=False)
+
+    def __init__(self, user_type):
+        self.user_type = user_type
