+++ api/model/database/users.py
@@ -28,14 +28,6 @@ class Usuario(db.Model):
         self.data_registro = datetime.datetime.now()
         self.bairro = bairro
 
-class Privilegio(db.Model):
-    __tablename__ = 'privilegios'
-    id = db.Column(db.Integer, primary_key=True)
-    user_type = db.Column(db.String(80), unique=True, nullable=False)
-
-    def __init__(self, user_type):
-        self.user_type = user_type
-
 class Bairro(db.Model):
     __tablename__ = 'bairros'
     id = db.Column(db.Integer, primary_key=True)
